Separate Bits from Bitfield
